Add i18n module to handle Apps translations
